Modernize SiPixel2DTemplateDBObjectESProducer
- move contents of header to source file
- use ESGetToken
